modified app.jsx and index.html
